Switch to 64-bit counters for rows read/written
The size of "usize" depends on the platform. On a 32-bit machine, the
maximum number is 4294967295, which can easily be overflown with enough
sustained throughput. Let's switch to explicit 64-bit counters, which
means the counter never wraps around in practice.

+++ libsql-server/sqld/src/database/libsql.rs
@@ -332,9 +332,9 @@ impl Connection {
 
     fn update_stats(&self, stmt: &rusqlite::Statement) {
         self.stats
-            .inc_rows_read(stmt.get_status(StatementStatus::RowsRead) as usize);
+            .inc_rows_read(stmt.get_status(StatementStatus::RowsRead) as u64);
         self.stats
-            .inc_rows_written(stmt.get_status(StatementStatus::RowsWritten) as usize);
+            .inc_rows_written(stmt.get_status(StatementStatus::RowsWritten) as u64);
     }
 }
